requirements.txt
aiohttp==3.7.4.post0
article==0.1.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
chardet==4.0.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap-v5==1.0.5
django-on-heroku==1.1.2
django-registration==3.2
django-tinymce==3.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
multidict==5.2.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pystache==0.5.4
python-dotenv==0.19.0
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
subject==0.1.0
typing-extensions==3.10.0.2
whitenoise==5.3.0
yarl==1.7.0
